smart-city-backend/requirements.txt
fastapi
uvicorn
httpx
python-dotenv
pydantic
openai
jinja2
weasyprint
